requirements.txt
spotipy
gin-config
